requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.2.1
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.58.0
typing-extensions==3.7.4.3
